factor out metacontroller approach  (#1007)
